+++ HISTORY.md
@@ -10,6 +10,7 @@
 7. Click on pair assigns correct message 
 8. Clicking on history item assigns message
 9. Save history on close
+10. Delete item from history
 
 
 ## Version 0.0.8 01 September 2026